glz history edit

+++ jp/admin/includes/functions/general.php
@@ -3189,7 +3189,7 @@ function tep_get_kakuukosuu_by_products_id($products_id) {
 function tep_get_kakaku_by_products_id($categories_id, $products_id){
   $data = tep_db_fetch_array(tep_db_query("select * from set_menu_list where categories_id='".$categories_id."' and products_id='".$products_id."'"));
   if ($data) {
-    return (int)$data['kakaku'];
+    return (float)$data['kakaku'];
   } else {
     return 0;
   }
